fix(deps): use path dependency for brick_core in brick_offline_first
The `brick_offline_first` package was depending on `brick_core` from `hosted`
which was causing a dependency conflict with other `brick` packages that
were using a path dependency for `brick_core`.

This change updates the `pubspec.yaml` of `brick_offline_first` to use a
path dependency for `brick_core` and `brick_sqlite` to resolve the conflict.