fix(deps): use path dependency for brick_core in brick_rest
The `brick_rest` package was depending on `brick_core` from `hosted`
which was causing a dependency conflict with other `brick` packages that
were using a path dependency for `brick_core`.

This change updates the `pubspec.yaml` of `brick_rest` to use a
path dependency for `brick_core` to resolve the conflict.