fix(deps): use path dependencies in brick_rest_generators
The `brick_rest_generators` package was depending on several `brick`
packages from `hosted` which was causing a dependency conflict with other
`brick` packages that were using path dependencies.

This change updates the `pubspec.yaml` of `brick_rest_generators` to
use path dependencies for all `brick` packages to resolve the conflict.